Move relevance computation to new source file.
